Implement "setup" command
Setup command which writes a file named "nimble.paths" with file system
paths to the package dependencies is implemented. The command also adds
code to include the paths file in "config.nims" configuration file to
make it available for the compiler. If "config.nims" does not exist it
will be created.

Related to nim-lang/nimble#127